Update cli_run cli_args test
Note: this answer is less correct,
but that is because zigs sub answers are more correct.

Zig specifically gives a more accurate answer for Num.log 3.
old answer:     1.0986122886681098
zig's answer:   1.0986122886681096
correct answer: 1.098612288668109691395

Both answers are technically correct when limited to doubles.
Zig's answer is closer to truly correct.
That said, the more correct answer leads to a slight rounding error in the final division.
So the new output is 4.000000000000001 instead of 4.0