Give up on exhaustively enumerating aliases
Keep getting errors in the build ruby example on
Json, and it doesn't tell me the name of the
symbol, just the IdentId, which makes this
really hard to track down.

This approach works fine, it'll just be harder
to debug the next time we run into a special
case like ElemWalker.